+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0]
+
 ### Added
 
 - Several internal types are now exported
@@ -103,6 +105,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Vahv itself! This is its first release.
 
 [Unreleased]: https://github.com/unleashy/vahv/compare/v0.4.0...HEAD
+[0.5.0]: https://www.npmjs.com/package/vahv/v/0.5.0
 [0.4.0]: https://www.npmjs.com/package/vahv/v/0.4.0
 [0.3.2]: https://www.npmjs.com/package/vahv/v/0.3.2
 [0.3.1]: https://www.npmjs.com/package/vahv/v/0.3.1